Resolve code comments
