+++ fe/src/assets/css/tailwind.css
@@ -4,43 +4,32 @@
 @tailwind components;
 @tailwind utilities;
 
-:root {
-  --gradient: linear-gradient(180deg, #f39d40 0%, #ac3971 100%);
-  --primary-50: "#F9F8F8";
-  --hr-gradient: linear-gradient(90deg, #cf7a1e 0%, #ac3971 100%);
-}
-
 body {
-  background: var(--primary-50);
   font-family: "Space Mono", monospace;
-  @apply text-primary-10;
+  @apply text-primary-10 bg-primary-50;
 }
 
-.btn-gradient {
-  background: var(--gradient);
-  @apply h-10 px-4 rounded-[6px] text-white;
+@layer base {
+  .bg-linear-gradient-primary {
+    @apply bg-gradient-to-r from-[#CF7A1E] to-[#AC3971];
+  }
+
+  .bg-linear-gradient-white-to-light {
+    @apply bg-gradient-to-b from-white to-light;
+  }
 }
 
-.btn-gradient-hr {
-  background: var(--hr-gradient);
-  @apply py-2 px-4 rounded-lg text-white text-sm font-bold;
-}
+@layer components {
+  .btn-gradient {
+    @apply h-10 px-4 rounded-[6px] text-white bg-gradient;
+  }
 
-.dropdown-item {
-  @apply p-2 text-left hover:bg-gray-200;
-}
+  .btn-gradient-hr {
+    background: var(--hr-gradient);
+    @apply py-2 px-4 rounded-lg text-white text-sm font-bold bg-linear-gradient-primary;
+  }
 
-.bg-linear-gradient-primary {
-  @apply bg-gradient-to-r from-[#CF7A1E] to-[#AC3971];
-}
-
-.bg-linear-gradient-white-to-light {
-  @apply bg-gradient-to-b from-white to-light;
-}
-
-.truncate {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  width: 100%;
+  .dropdown-item {
+    @apply p-2 text-left hover:bg-gray-200;
+  }
 }
