+++ fe/src/components/base/dropdown.vue
@@ -9,7 +9,7 @@
         </MenuButton>
       </div>
 
-      <transition
+      <Transition
         enter-active-class="transition duration-100 ease-out"
         enter-from-class="transform scale-95 opacity-0"
         enter-to-class="transform scale-100 opacity-100"
@@ -78,7 +78,7 @@
             </MenuItem>
           </div>
         </MenuItems>
-      </transition>
+      </Transition>
     </Menu>
   </div>
 </template>
