+++ fe/src/components/base/search-bar.vue
