+++ fe/src/components/campaign/campaign-card.vue
