+++ fe/src/components/global/navbar.vue
@@ -1,15 +1,13 @@
 <template>
   <header class="shadow-sm bg-white/30 h-16 py-3 px-16">
     <nav class="flex justify-between items-center">
-      <NuxtLink to="/" class="justify-center"><Logo /></NuxtLink>
+      <NuxtLink to="/" class="justify-center"><IconsLogo /></NuxtLink>
       <div class="flex gap-4">
-        <BaseSearch />
+        <BaseSearchBar />
         <WalletButton />
       </div>
     </nav>
   </header>
 </template>
 
-<script setup lang="ts">
-import Logo from "../components/icons/Logo.vue";
-</script>
+<script setup lang="ts"></script>
