+++ fe/src/components/icons/blocks-right.vue
