+++ fe/src/components/user/user-card.vue
@@ -3,7 +3,7 @@
     class="rounded-2xl w-80 h-40 bg-gradient-to-b from-[#F39D40] to-[#AC3971] p-[1px]"
   >
     <div class="bg-white h-full rounded-2xl p-6 space-x-4 flex items-center">
-      <img :src="imgSrc" class="h-24 w-24 rounded-full object-cover" />
+      <UserAvatar :img-src="imgSrc" :height="90" :width="90" />
       <div>
         <div class="flex text-2xl font-bold space-x-4">
           <div class="text-dark">{{ blockChainValue }}</div>
