+++ fe/src/composables/useUtils.ts
@@ -1,5 +1,5 @@
 export const useUtils = () => {
-  const truncate = (text: string, length: number) => {
+  const truncate = (text: string, length: number): string => {
     if (text.length > length) {
       return text.substring(0, length) + "...";
     } else {
@@ -11,7 +11,7 @@ export const useUtils = () => {
     address: string,
     startLength: number,
     endLength: number
-  ) => {
+  ): string => {
     if (address.length <= startLength + endLength) {
       return address;
     }
