+++ fe/src/mocks/campaign-sample.json
@@ -9,7 +9,7 @@
   "imgSrc": "https://images.unsplash.com/photo-1529390079861-591de354faf5?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1170&q=80",
   "story": "Lorem ipsum dolor sit amet consectetur. Amet senectus augue in urna. Lorem ipsum dolor sit amet consectetur. Amet senectus augue in urna.Lorem ipsum dolor sit amet consectetur. Amet senectus augue in urna. Lorem ipsum dolor sit amet consectetur. Amet senectus augue in urna.Lorem ipsum dolor sit amet consectetur. Amet senectus augue in urna.",
   "daysLeft": 190,
-  "donations": "0.23",
+  "donations": 0.23,
   "totalBackers": 23,
   "donators": [
     {
