+++ fe/src/pages/create-campaign.vue
@@ -43,7 +43,7 @@
             :model-value="story"
             @change="handleChange"
           />
-          <FileUpload />
+          <BaseFileUpload />
           <div class="flex w-full space-x-4">
             <BaseInput
               id="goal"
@@ -159,7 +159,7 @@ const onSubmit = handleSubmit(async () => {
         values.campaign,
         values.story,
         ethers.parseEther((values.goal as number).toString()),
-        deadline.getTime() / 1000, // convert from milliseconds to secsonds
+        deadline.getTime() / 1000 // convert from milliseconds to secsonds
       )
       .then(() => {
         resetForm();
